requirements.txt
absl-py==2.4.0
accelerate==1.12.0
aiofiles==23.2.1
alembic==1.18.4
altair==6.0.0
altgraph==0.17.5
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
astunparse==1.6.3
async-timeout==5.0.1
asyncpg==0.31.0
attrs==25.4.0
backports.asyncio.runner==1.2.0
bcrypt==4.0.1
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==7.0.3
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.10.1
contourpy==1.3.2
cryptography==46.0.5
cycler==0.12.1
deepsearch-glm==0.22.0
defusedxml==0.7.1
dill==0.4.1
dnspython==2.8.0
docling==2.73.1
docling-core==2.65.1
docling-ibm-models==3.11.0
docling-parse==4.7.3
docutils==0.22.4
ecdsa==0.19.1
email-validator==2.3.0
et_xmlfile==2.0.0
exceptiongroup==1.3.1
Faker==40.4.0
fastapi==0.129.0
filelock==3.24.2
filetype==1.2.0
Flask==3.1.2
flatbuffers==25.12.19
fonttools==4.61.1
fsspec==2026.2.0
gast==0.7.0
gitdb==4.0.12
GitPython==3.1.46
gliner==0.2.26
google-auth==2.48.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==3.3.1
grpcio==1.78.0
h11==0.16.0
h5py==3.15.1
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.26.0
huggingface_hub==0.36.2
humanfriendly==10.0
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
jsonlines==4.0.0
jsonref==1.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
keras==3.12.1
Keras-Preprocessing==1.1.2
kiwisolver==1.4.9
langdetect==1.0.9
latex2mathml==3.78.1
libclang==18.1.1
lxml==6.0.2
Mako==1.3.10
Markdown==3.10.2
markdown-it-py==4.0.0
marko==2.2.2
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
ml_dtypes==0.5.4
mpire==2.10.2
mpmath==1.3.0
multiprocess==0.70.19
namex==0.1.0
narwhals==2.17.0
networkx==3.4.2
netwulf==0.1.5
numerize==0.12
numpy==1.26.4
oauthlib==3.3.1
omegaconf==2.3.0
onnxruntime==1.23.2
opencv-python==4.11.0.86
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.18.0
packaging==26.0
pandas==2.3.3
passlib==1.7.4
pefile==2024.8.26
pillow==12.1.1
pipdeptree==2.30.0
pluggy==1.6.0
polyfactory==3.2.0
protobuf==6.33.5
psutil==7.2.2
psycopg2-binary==2.9.11
pyaes==1.6.1
pyarrow==23.0.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pyclipper==1.4.0
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.0
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pyinstaller==6.19.0
pyinstaller-hooks-contrib==2026.0
pylatexenc==2.10
PyMuPDF==1.27.1
pyparsing==3.3.2
pypdfium2==5.4.0
pyreadline3==3.5.4
Pyrogram==2.0.106
PySocks==1.7.1
pytest==9.0.2
pytest-asyncio==1.3.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.22
python-pptx==1.0.2
pytz==2025.2
pywifi==1.1.12
pywin32==306
pywin32-ctypes==0.2.3
PyYAML==6.0.3
rapidocr==3.6.0
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.30.0
rsa==4.9.1
rtree==1.4.1
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
semchunk==2.2.2
sentence-transformers==5.1.2
sentencepiece==0.2.1
shapely==2.1.2
shellingham==1.5.4
simplejson==3.20.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8.3
spark-parser==1.9.0
SQLAlchemy==2.0.46
starlette==0.52.1
streamlit==1.55.0
sympy==1.14.0
tabulate==0.9.0
Telethon==1.42.0
tenacity==9.1.4
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow==2.20.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.3.0
tf_keras==2.20.1
TgCrypto==1.2.5
threadpoolctl==3.6.0
tokenizers==0.22.2
toml==0.10.2
tomli==2.4.0
torch==2.10.0
torchvision==0.25.0
tornado==6.5.4
tqdm==4.67.3
transformers==4.57.6
tree-sitter==0.25.2
tree-sitter-c==0.24.1
tree-sitter-javascript==0.25.0
tree-sitter-python==0.25.0
tree-sitter-typescript==0.23.2
typer==0.21.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uncompyle6==3.9.3
unpy2exe==0.4
urllib3==2.6.3
uvicorn==0.27.1
watchdog==6.0.0
watchfiles==1.1.1
websockets==16.0
Werkzeug==3.1.5
wrapt==2.1.1
xdis==6.1.8
xlsxwriter==3.2.9
yt-dlp==2026.2.4
